requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
pydantic==2.8.2
python-dotenv==1.0.1

Pillow==10.4.0
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.10

openai==1.40.6
pyspellchecker==0.8.1
textstat==0.7.4
reportlab==4.2.2
